+++ app/controllers/api_controller.rb
@@ -1,0 +1,4 @@
+class ApiController < ActionController::Base
+  include DeviseTokenAuth::Concerns::SetUserByToken
+  protect_from_forgery with: :null_session
+end
