+++ config/routes.rb
@@ -1,6 +1,8 @@
 Rails.application.routes.draw do
-  resources :books, only: [:index, :show]
-
+  api_version(module: 'api/v1', path: { value: 'api/v1' }, defaults: { format: :json }) do
+    resources :books, only: [:index, :show]
+  end
+  
   mount_devise_token_auth_for 'User', at: 'auth'
   # For details on the DSL available within this file, see http://guides.rubyonrails.org/routing.html
 end
