+++ spec/controllers/books_controller_spec.rb
@@ -1,4 +1,4 @@
 require 'rails_helper'
 
-describe BooksController do
+describe Api::V1::BooksController do
 end
